requirements.txt
rembg[cpu]==2.0.57
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0